Code can now compiled with -Wconversion

+++ lib-artnet/src/artnetnodehandledmxin.cpp
@@ -26,8 +26,8 @@
  * THE SOFTWARE.
  */
 
-#include <stdint.h>
-#include <string.h>
+#include <cstdint>
+#include <cstring>
 #include <cassert>
 
 #include "artnetnode.h"
@@ -41,7 +41,7 @@
 
 static uint32_t s_ReceivingMask = 0;
 
-void ArtNetNode::SetDestinationIp(uint8_t nPortIndex, uint32_t nDestinationIp) {
+void ArtNetNode::SetDestinationIp(uint32_t nPortIndex, uint32_t nDestinationIp) {
 	if (nPortIndex < ARTNET_NODE_MAX_PORTS_INPUT) {
 		if (Network::Get()->IsValidIp(nDestinationIp)) {
 			m_InputPorts[nPortIndex].nDestinationIp = nDestinationIp;
@@ -65,15 +65,15 @@ void ArtNetNode::HandleDmxIn() {
 		uint32_t nUpdatesPerSecond;
 
 		if (m_InputPorts[i].bIsEnabled){
-			uint16_t nLength;
+			uint32_t nLength;
 			const auto *pDmxData = m_pArtNetDmx->Handler(i, nLength, nUpdatesPerSecond);
 
 			if (pDmxData != nullptr) {
 				tArtDmx.Sequence = 1 + m_InputPorts[i].nSequence++;
-				tArtDmx.Physical = i;
+				tArtDmx.Physical = static_cast<uint8_t>(i);
 				tArtDmx.PortAddress = m_InputPorts[i].port.nPortAddress;
-				tArtDmx.LengthHi = (nLength & 0xFF00) >> 8;
-				tArtDmx.Length = (nLength & 0xFF);
+				tArtDmx.LengthHi = static_cast<uint8_t>((nLength & 0xFF00) >> 8);
+				tArtDmx.Length = static_cast<uint8_t>(nLength & 0xFF);
 
 				memcpy(tArtDmx.Data, pDmxData, nLength);
 
@@ -86,7 +86,7 @@ void ArtNetNode::HandleDmxIn() {
 			} else {
 				if ((m_InputPorts[i].port.nStatus & GO_DATA_IS_BEING_TRANSMITTED) == GO_DATA_IS_BEING_TRANSMITTED) {
 					if (nUpdatesPerSecond == 0) {
-						m_InputPorts[i].port.nStatus = m_InputPorts[i].port.nStatus & ~GI_DATA_RECIEVED;
+						m_InputPorts[i].port.nStatus = static_cast<uint8_t>(m_InputPorts[i].port.nStatus & ~GI_DATA_RECIEVED);
 						s_ReceivingMask &= ~(1U << i);
 						m_State.bIsReceivingDmx = (s_ReceivingMask != 0);
 					}
