+++ lib-dmx/src/rpi/dmx++.cpp
@@ -23,7 +23,7 @@
  * THE SOFTWARE.
  */
 
-#include <stdint.h>
+#include <cstdint>
 #include <cassert>
 
 #include "dmx.h"
@@ -32,14 +32,14 @@
 #include "arm/pl011.h"
 #include "bcm2835.h"
 
-const uint8_t *Dmx::RdmReceive(__attribute__((unused)) uint8_t nPort) {
+const uint8_t *Dmx::RdmReceive(__attribute__((unused)) uint32_t nPort) {
 	assert(nPort == 0);
 
 	const auto *p = rdm_get_available();
 	return p;
 }
 
-const uint8_t* Dmx::RdmReceiveTimeOut(uint8_t nPort, uint32_t nTimeOut) {
+const uint8_t* Dmx::RdmReceiveTimeOut(uint32_t nPort, uint32_t nTimeOut) {
 	assert(nPort == 0);
 
 	uint8_t *p = nullptr;
@@ -54,13 +54,13 @@ const uint8_t* Dmx::RdmReceiveTimeOut(uint8_t nPort, uint32_t nTimeOut) {
 	return p;
 }
 
-void Dmx::SetPortDirection(__attribute__((unused)) uint8_t nPort, TDmxRdmPortDirection tPortDirection, bool bEnableData) {
+void Dmx::SetPortDirection(__attribute__((unused)) uint32_t nPort, TDmxRdmPortDirection tPortDirection, bool bEnableData) {
 	assert(nPort == 0);
 
 	dmx_set_port_direction(static_cast<_dmx_port_direction>(tPortDirection), bEnableData);
 }
 
-void Dmx::RdmSendRaw(__attribute__((unused)) uint8_t nPort, const uint8_t *pRdmData, uint16_t nLength) {
+void Dmx::RdmSendRaw(__attribute__((unused)) uint32_t nPort, const uint8_t *pRdmData, uint16_t nLength) {
 	assert(nPort == 0);
 
 	BCM2835_PL011->LCRH &= ~PL011_LCRH_FEN;
